@@ -2,33 +2,40 @@
 #include <iostream>
 #include <limits>
 #include <time.h>
-#include <float.h>
 using namespace std;
+
 class AssignmentProblemSolver
 {
 private:
     // --------------------------------------------------------------------------
     // Computes the optimal assignment (minimum overall costs) using Munkres algorithm.
     // --------------------------------------------------------------------------
-    void assignmentoptimal( int *assignment, double *cost, double *distMatrix, int nOfRows, int nOfColumns );
-    void buildassignmentvector( int *assignment, bool *starMatrix, int nOfRows, int nOfColumns );
-    void computeassignmentcost( int *assignment, double *cost, double *distMatrix, int nOfRows );
-    void step2a( int *assignment, double *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, int nOfRows, int nOfColumns, int minDim );
-    void step2b( int *assignment, double *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, int nOfRows, int nOfColumns, int minDim );
-    void step3 ( int *assignment, double *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, int nOfRows, int nOfColumns, int minDim );
-    void step4 ( int *assignment, double *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, int nOfRows, int nOfColumns, int minDim, int row, int col );
-    void step5 ( int *assignment, double *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, int nOfRows, int nOfColumns, int minDim );
+    void assignmentoptimal( vector<int> &assignment, float &cost, const vector<float> &distMatrixIn, size_t nOfRows, size_t nOfColumns );
+    void buildassignmentvector( vector<int> &assignment, bool *starMatrix, size_t nOfRows, size_t nOfColumns );
+    void computeassignmentcost( const vector<int> &assignment, float &cost, const vector<float> &distMatrixIn, size_t nOfRows );
+    void step2a( vector<int> &assignment, float *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, size_t nOfRows, size_t nOfColumns, size_t minDim );
+    void step2b( vector<int> &assignment, float *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, size_t nOfRows, size_t nOfColumns, size_t minDim );
+    void step3( vector<int> &assignment, float *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, size_t nOfRows, size_t nOfColumns, size_t minDim );
+    void step4( vector<int> &assignment, float *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, size_t nOfRows, size_t nOfColumns, size_t minDim, size_t row, size_t col );
+    void step5( vector<int> &assignment, float *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, size_t nOfRows, size_t nOfColumns, size_t minDim );
     // --------------------------------------------------------------------------
     // Computes a suboptimal solution. Good for cases with many forbidden assignments.
     // --------------------------------------------------------------------------
-    void assignmentsuboptimal1( int *assignment, double *cost, double *distMatrixIn, int nOfRows, int nOfColumns );
+    void assignmentsuboptimal1( vector<int> &assignment, float &cost, const vector<float> &distMatrixIn, size_t nOfRows, size_t nOfColumns );
     // --------------------------------------------------------------------------
     // Computes a suboptimal solution. Good for cases with many forbidden assignments.
     // --------------------------------------------------------------------------
-    void assignmentsuboptimal2( int *assignment, double *cost, double *distMatrixIn, int nOfRows, int nOfColumns );
+    void assignmentsuboptimal2( vector<int> &assignment, float &cost, const vector<float> &distMatrixIn, size_t nOfRows, size_t nOfColumns );
+
 public:
-    enum TMethod { optimal, many_forbidden_assignments, without_forbidden_assignments };
+    enum TMethod
+    {
+        optimal,
+        many_forbidden_assignments,
+        without_forbidden_assignments
+    };
+
     AssignmentProblemSolver();
     ~AssignmentProblemSolver();
-    double Solve( vector<vector<double>> &DistMatrix, vector<int> &Assignment, TMethod Method = optimal );
+    float Solve( const vector<float> &distMatrixIn, size_t nOfRows, size_t nOfColumns, vector<int> &assignment, TMethod Method = optimal );
 };
